Transaction screens for category and account

+++ lib/models/category_model.dart
@@ -1,6 +1,6 @@
 class TCategory {
   String name;
-  int monthlyBudget;
+  double monthlyBudget;
   String uid;
 
   static List<TCategory> dummyCategories = [
